Make the isSqlEQ matcher more official
