+++ packages/malloy-db-test/src/expr.spec.ts
@@ -15,33 +15,9 @@
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 
 import * as malloy from "@malloydata/malloy";
-import { Result } from "@malloydata/malloy";
 import { RuntimeList } from "./runtimes";
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace jest {
-    interface Matchers<R> {
-      isSqlEq(): R;
-    }
-  }
-}
-
-expect.extend({
-  isSqlEq: (result: Result) => {
-    const wantEq = result.data.path(0, "calc").value;
-    if (wantEq != "=") {
-      return {
-        pass: false,
-        message: () => `${wantEq}\nSQL:\n${result.sql}`,
-      };
-    }
-    return {
-      pass: true,
-      message: () => "SQL expression matched",
-    };
-  },
-});
+import "./sql-eq";
+import { mkSqlEq } from "./sql-eq";
 
 const runtimes = new RuntimeList([
   "bigquery", //
@@ -595,43 +571,14 @@ expressionModels.forEach((expressionModel, databaseName) => {
 });
 
 runtimes.runtimeMap.forEach((runtime, databaseName) => {
-  async function sqlEq(expr: string, result: string | boolean) {
-    const qExpr = expr.replace(/'/g, "`");
-    const sourceDef = `
-      sql: nothing is || SELECT 1 as one ;;
-      source: basicTypes is from_sql(nothing) + {
-        dimension: friName is 'friday'
-        dimension: friDay is 5
-        dimension: satName is 'saturday'
-        dimension: satDay is 6
-        dimension: dayTime is @2001-01-01 00:00:00
-      }\n`;
-    let query: string;
-    if (typeof result == "boolean") {
-      const notEq = `concat('${qExpr} was ${!result} expected ${result}')`;
-      const whenPick = result ? "'=' when exprTrue" : `${notEq} when exprTrue`;
-      const elsePick = result ? notEq : "'='";
-      query = `${sourceDef}
-        query: basicTypes
-        -> { project: exprTrue is ${expr} }
-        -> {
-          project: calc is pick ${whenPick} else ${elsePick}
-        }`;
-    } else {
-      const qResult = result.replace(/'/g, "`");
-      query = `${sourceDef}
-        query: basicTypes
-        -> {
-          project: expect is ${result}
-          project: got is ${expr}
-        } -> {
-          project: calc is
-            pick '=' when expect = got
-            else concat('${qExpr} != ${qResult}. Got: ', got::string)
-        }`;
-    }
-    return await runtime.loadQuery(query).run();
-  }
+  const sqlEq = mkSqlEq(runtime, {
+    malloy: `+ {
+      dimension: friName is 'friday'
+      dimension: friDay is 5
+      dimension: satName is 'saturday'
+      dimension: satDay is 6
+    }`,
+  });
 
   describe.skip(`alternations with not-eq - ${databaseName}`, () => {
     /*
